WIP: TW-323: RPC list altering update. Refactor
