+++ src/modals/custom-rpc-modals/form.utils.ts
@@ -3,12 +3,17 @@ import { object, SchemaOf, string } from 'yup';
 import { makeRequiredErrorMessage } from 'src/form/validation/messages';
 import { urlValidation } from 'src/form/validation/url';
 
-type AddCustomRpcFormValues = {
+interface FormValues {
   name: string;
   url: string;
+}
+
+export const addCustomRpcFormInitialValues: FormValues = {
+  name: '',
+  url: ''
 };
 
-export const editCustomRpcFormValidationSchema: SchemaOf<AddCustomRpcFormValues> = object().shape({
+export const formValidationSchema: SchemaOf<FormValues> = object().shape({
   name: string().required(makeRequiredErrorMessage('Name')),
   url: urlValidation
 });
